move hasMany into model, no attempt at tests passing yet
