+++ packages/store/addon/-private/system/references.ts
@@ -2,7 +2,6 @@
   @module @ember-data/store
 */
 
-import HasManyReference from './references/has-many';
 import RecordReference from './references/record';
 
-export { RecordReference, HasManyReference };
+export { RecordReference };
